backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<26.0
drf-spectacular>=0.27,<1.0
Pillow>=10.0,<12.0
psycopg2-binary>=2.9,<3.0
cloudinary>=1.39,<2.0
django-cloudinary-storage>=0.3,<1.0
dj-database-url>=2.2,<3.0
django-cors-headers>=4.4,<5.0
whitenoise>=6.7,<7.0
gunicorn>=22.0,<24.0
